Carry explicit encoding context along as we encode values
This allows us to correctly encode values with different indentation
in lists and records. As a side-effect of this change, the way
indentation is rendered has changed such that there are `indent - 1`
spaces between multi-line list markers and list elements now. I believe
this is most consistent.

This commit adds tests to ensure that encoding of records in
lists (both inline and not inline) are correct.

Closes #29